pyproject.toml
[tool.poetry]
name = "rew"
version = "0.2.1"
description = "to quickly rewrite things"
authors = ["Vinay Sheel Sagar <vinaysheelsagar@gmail.com>"]
license = "mit"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.9"
google-generativeai = "^0.3.2"
pyyaml = "^6.0.1"
click = "^8.1.7"


[tool.poetry.scripts]
rew = "rew.__main__:gateway"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
